requirements.txt
psycopg2==2.6.1
dj-database-url==0.4.1
whitenoise==3.0
gunicorn==19.4.1
djangorestframework==3.3.3
Django==1.9.4
